+++ examples/2d/plot_mnist_classify_torch.py
@@ -49,7 +49,7 @@ train_data.targets = train_data.targets[random_permute]
 train_loader = torch.utils.data.DataLoader(train_data,
     batch_size=128, shuffle=True, num_workers=num_workers, pin_memory=pin_memory)
 
-#Creat the test loader on the full MNIST test set
+#Creat the tests loader on the full MNIST tests set
 test_loader = torch.utils.data.DataLoader(
     datasets.MNIST(
         scattering_datasets.get_dataset_dir('MNIST'),
@@ -73,7 +73,7 @@ class View(nn.Module):
 
 
 ###############################################################################
-# Create a training and test function
+# Create a training and tests function
 import torch.nn.functional as F
 
 def train(model, device, train_loader, optimizer, scattering):
@@ -127,4 +127,4 @@ for epoch in range(0, 20):
     train( model, device, train_loader, optimizer, scattering)
 
 acc = test(model, device, test_loader, scattering)
-print('Scattering order  2 linear model test accuracy: %.2f'%(acc) )
+print('Scattering order  2 linear model tests accuracy: %.2f'%(acc) )
