+++ kymatio/backend/numpy_backend.py
@@ -18,3 +18,10 @@ def modulus(x):
 
     """
     return np.abs(x)
+
+def _iscomplex(x):
+    return x.dtype == np.complex64 or x.dtype == np.complex128
+
+
+def _isreal(x):
+    return x.dtype == np.float32 or x.dtype == np.float64
